+++ AllTopLevelVoiceAppsToMarkdownDocFx.ps1
@@ -6,11 +6,17 @@
     This script can only be used when it's called from another script in another repo. Do not use it manually.
 
     Author:             Martin Heusser
-    Version:            1.0.4
+    Version:            1.0.5
 
 #>
 
 #Requires -Modules @{ ModuleName = "MicrosoftTeams"; ModuleVersion = "5.0.0" }, "Microsoft.Graph.Users", "Microsoft.Graph.Groups"
+
+[CmdletBinding(DefaultParametersetName="None")]
+param(
+    [Parameter(Mandatory=$true)][String]$ArticlesRelativePath,
+    [Parameter(Mandatory=$false)][Switch]$IncludeCallFlowDocs
+)
 
 # Import external functions via dot sourcing
 . .\Functions\Connect-M365CFV.ps1
@@ -59,14 +65,22 @@ foreach ($CallQueue in $AllCallQueues) {
     
 }
 
-Set-Content -Path "$localRepoPath\Docs\articles\callflows\callflows.md" -Value "# Call Flows"
+Set-Content -Path "$localRepoPath\Docs\articles\$ArticlesRelativePath\call_flows.md" -Value "# Call Flows"
+
+if ($IncludeCallFlowDocs) {
+
+    Add-Content -Path "$localRepoPath\Docs\articles\$ArticlesRelativePath\call_flows.md" -Value "`n[!INCLUDE[](call_flow_docs.md)]`n"
+
+}
+
+$VoiceApps = $VoiceApps | Sort-Object -Unique
 
 foreach ($VoiceAppIdentity in $VoiceApps) {
 
-    . .\M365CallFlowVisualizerV2.ps1 -Identity $VoiceAppIdentity -Theme dark -CustomFilePath "$localRepoPath\Docs\articles\callflows\$voiceAppIdentity" -ShowCqAgentPhoneNumbers -ExportAudioFiles -ExportTTSGreetings -ShowAudioFileName -ShowTTSGreetingText -ExportPng $true -CacheResults $true -ExportHtml $true -DocFxMode
+    . .\M365CallFlowVisualizerV2.ps1 -Identity $VoiceAppIdentity -Theme dark -CustomFilePath "$localRepoPath\Docs\articles\$ArticlesRelativePath\$voiceAppIdentity" -ShowCqAgentPhoneNumbers -ExportAudioFiles -ExportTTSGreetings -ShowAudioFileName -ShowTTSGreetingText -ExportPng $true -CacheResults $true -ExportHtml $true -DocFxMode -ShowCqAgentOptInStatus -ShowSharedVoicemailGroupMembers $true
 
     $markdownInclude = "&nbsp;`n[!include[$($VoiceAppFileName)]($voiceAppIdentity/$(($VoiceAppFileName).Replace(" ","_"))_CallFlow$fileExtension)]`n- [Enlarge View]($voiceAppIdentity/$(($VoiceAppFileName).Replace(" ","_"))_CallFlow.htm)`n- [PNG Download]($voiceAppIdentity/$(($VoiceAppFileName).Replace(" ","_"))_CallFlow.png) `n&nbsp;`n"
 
-    Add-Content -Path "$localRepoPath\Docs\articles\callflows\callflows.md" -Value $markdownInclude
+    Add-Content -Path "$localRepoPath\Docs\articles\$ArticlesRelativePath\call_flows.md" -Value $markdownInclude
 
 }