+++ FindUnusedVoiceAppsAndResourceAccounts.ps1
@@ -17,18 +17,12 @@ param(
 )
 
 . .\Functions\Connect-M365CFV.ps1
+. .\Functions\Get-AllVoiceAppsAndResourceAccounts.ps1
 
 . Connect-M365CFV
 
-# Read all Auto Attendants, Call Queues and Resource Accounts into memory
-Write-Host "Retrieving all Auto Attendants (max. 1000)... this can take a while..." -ForegroundColor Magenta
-$global:allAutoAttendants = Get-CsAutoAttendant -First 1000
-
-Write-Host "Retrieving all Call Queues (max. 1000)... this can take a while..." -ForegroundColor Magenta
-$global:allCallQueues = Get-CsCallQueue -WarningAction SilentlyContinue -First 1000
-
-Write-Host "Retrieving all Resource Accounts (max. 1000)... this can take a while..." -ForegroundColor Magenta
-$global:allResourceAccounts = Get-CsOnlineApplicationInstance -ResultSize 1000
+# Get all voice apps and resource accounts from external function
+. Get-AllVoiceAppsAndResourceAccounts
 
 $report = @()